Moved `print_item` and helpers to a separate file
